requirements.txt
gradio
uvicorn
pydantic
python-dotenv
llama-cpp-python
langgraph
langchain_community
langchain_huggingface
sentence-transformers